various cleanups and project structure
* CS1591 disabled for the entire solution
* removed `[assembly: System.CLSCompliant(true)]` from contributed
projects
* Fixed various async snafus.
