+++ src/OpenTelemetry/Trace/Export/SpanExporter.cs
@@ -53,9 +53,7 @@ namespace OpenTelemetry.Trace.Export
         /// <inheritdoc/>
         public override Task ExportAsync(SpanData export, CancellationToken token)
         {
-            this.worker.ExportAsync(export, token);
-
-            return Task.CompletedTask;
+            return this.worker.ExportAsync(export, token);
         }
 
         public override void RegisterHandler(string name, IHandler handler)
